+++ TestProjects/Assets/app.ts
@@ -1,0 +1,3 @@
+import "lib";
+
+console.log("Nothing to see here");